+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 - GLS-173 - Add model + endpoint for handling UK total trade data by country
 - GLS-186 - Dataservices addition for market trends
+- GLS-186 - Dataservices addition for trade highlights
 
 ### Bugs fixed